credit_scroring_console/requirements.txt
email-validator==1.3.0
pydantic==1.10.2
psycopg2-binary==2.9.5
